Refactor JS models
