+++ android/src/main/java/com/sicpa/didcomm/reactnative/DIDDocResolverProxy.kt
@@ -20,7 +20,7 @@ class DIDDocResolverProxy(private val resolversProxyModule: ResolverProxyModule)
 
         runBlocking {
             try {
-                withTimeout(5000) {
+                withTimeout(1500) {
                     scope.launch {
                         resolversProxyModule.sendEvent(ResolveDid(did))
                         resolvedDid = resolvedDidChannel.receive()
