+++ android/src/main/java/com/sicpa/didcomm/reactnative/model/JSDIDCommService.kt
@@ -1,0 +1,9 @@
+package com.sicpa.didcomm.reactnative.model
+
+data class JSDIDCommService(val id: String, val kind: JSServiceKind)
+data class JSServiceKind(val DIDCommMessaging: JSDIDCommMessagingService)
+data class JSDIDCommMessagingService(
+    val service_endpoint: String,
+    val accept: List<String>,
+    val route_keys: List<String>
+)
